Create separate widget for auth dialog box
